requirements.txt
langchain-community
langchain-text-splitters
groq
streamlit
python-dotenv
pypdf
faiss-cpu
scikit-learn
openai
plotly
pandas
